+++ src/app/loadout/ingame/ingame-loadout-apply.ts
@@ -9,7 +9,11 @@ import { t } from 'app/i18next-t';
 import { inGameLoadoutNotification } from 'app/inventory/MoveNotifications';
 import { itemMoved } from 'app/inventory/actions';
 import { updateCharacters } from 'app/inventory/d2-stores';
-import { allItemsSelector, storesSelector } from 'app/inventory/selectors';
+import {
+  allItemsSelector,
+  createItemContextSelector,
+  storesSelector,
+} from 'app/inventory/selectors';
 import { getStore } from 'app/inventory/stores-helpers';
 import { itemMoveLoadout } from 'app/loadout-drawer/auto-loadouts';
 import { applyLoadout } from 'app/loadout-drawer/loadout-apply';
@@ -23,6 +27,7 @@ import { DimError } from 'app/utils/dim-error';
 import { reportException } from 'app/utils/exceptions';
 import { errorLog } from 'app/utils/log';
 import { PlatformErrorCodes } from 'bungie-api-ts/destiny2';
+import { useSelector } from 'react-redux';
 import { inGameLoadoutDeleted, inGameLoadoutUpdated } from './actions';
 import { getItemsFromInGameLoadout } from './ingame-loadout-utils';
 
@@ -31,6 +36,7 @@ import { getItemsFromInGameLoadout } from './ingame-loadout-utils';
  */
 export function applyInGameLoadout(loadout: InGameLoadout): ThunkResult {
   return async (dispatch, getState) => {
+    const itemCreationContext = useSelector(createItemContextSelector);
     const account = currentAccountSelector(getState())!;
     const stores = storesSelector(getState());
     const target = getStore(stores, loadout.characterId)!;
@@ -43,7 +49,11 @@ export function applyInGameLoadout(loadout: InGameLoadout): ThunkResult {
       await applyPromise;
 
       // Find each item, and update it to be equipped!
-      const items = getItemsFromInGameLoadout(loadout.items, allItemsSelector(getState()));
+      const items = getItemsFromInGameLoadout(
+        itemCreationContext,
+        loadout.items,
+        allItemsSelector(getState())
+      );
 
       for (const item of items) {
         // Update items to be equipped